Selfhost: compiling errno (#467)
Add support for the `errno` intrinsic, now that `uname` exists and
allows for runtime OS detection (since the mechanism for fetching
the errno value differs per-platform). Also, this fixes the libc
function `strerror` which incorrectly had the `String` return type
and it should have had `Pointer<Byte>` in order to be compatible with
C (I'm not really sure how this worked in the reference implementation).