Selfhost: compiling imports and exports
This first pass does two things. Firstly, it removes the need for a
designated "entry" module when compiling; it just uses the
dependency-ordered module list that's built during typechecking (by
sorting the TypedModules by `id` we essentially obtain a list of modules
in the order that they need to be evaluated for the program to work).
Secondly, it adds an optional TypedModule field onto the
TypedAstNodeKind.Identifier variant, which will be used in the
compilation step to determine whether the identifier is an exported
value or not. But for now, I add some additional error reporting to
raise a typechecker error if the user attempts to reassign an imported
variable (or an aliased imported value).