Selfhost: compiling imports/exports
Typechecker: within the typechecker, I needed to update the way in which
moduleIds are assigned to modules to ensure that, when sorted in
ascending order, the resulting array was a valid dependency graph (ie.
given any module in the array, all of its dependencies will have lower
index values than that module's index value). This required some work,
but nothing too groundbreaking. It did force me to rely less on magic
numbers, so I can no longer assume that moduleId 0 was the prelude, and
1 was the intrinsics; what I have here is actually a much nicer
solution, and much more scalable as the standard library grows. It did
mean I had to update a lot of the typechecker test output files...

Compiler: I had to make a minor change around the ways in which known
intrinsic structs/enums were detected, since I could no longer make
assumptions about the moduleId. Also, add support for reading from and
writing to exported variables by storing them as data; given that each
module is loaded in dependency-graph order, by the time a given exported
variable is read, it is guaranteed to have had its value stored in it.