Selfhost: compiling string interpolation (#469)
Actually, before I could write code to compile string interpolation, I
needed to fix it! I had apparently not noticed a bug in the lexer (!)
for string interpolation expressions, which resulted in incorrect
sub-expressions being output.
Once that was fixed, emitting the proper code for string interpolation
was straightforward since it's very similar to the existing code for
string concatenation. Interpolation is a more efficient form of
concatenation because it doesn't allocate intermediate strings for each
binary sub-expression of concatenation.
I also added builtin support for calling the `print` function, much like
how the `println` function is handled. I did this in my quest to be able
to compile the selfhosted compiler with itself, and I noticed this was
missing.