Selfhost: use u64 for enum variant ids
Previously the memory layout for a Container enum variant (eg. an enum
variant that held data, rather than just a numerical id) was a 4 byte id
field to represent the variant index, followed by an arbitrary amount of
data needed to hold all of the variant's fields. However, as I started
to compile and larger programs (like, the selfhosted lexer itself!) I
noticed weird unexpected memory issues.
This turned out to be the result of the garbage collector freeing the
chunks of memory that I had allocated to hold the values of these enum
variants (aka, the data that fell _after_ the initial 4-byte id header).
This is because of the way that the garbage collector works - it scans
the data structures allocated via GC_malloc (or it variants) and looks
for things that resemble pointers which are _also_ allocated via
GC_malloc. However, since container enum variants' allocated memory was
not aligned on 8-byte divisions, the GC was not able to recognize that
the data in the value section of a container enum variant were
reachable, and they'd be freed and reused for future allocations.
By switching to a (slightly more wasteful) 8-byte header, the GC can
properly scan the data allocated for these data structures, and
determine the reachability of the internal allocated memory.