Selfhost: compiling lexer with selfhosted compiler!
I successfully compiled the selfhosted lexer (specifically the
`src/lexer.test.abra` file) and ran it with the test files in
`test/lexer/` to verify that the output was consistent when compared
with the output of running `src/lexer.test.abra` as compiled by the
reference compiler. In order to do this, I needed to refactor the test
runner a bit (and add some test output files which were previously
missing since it had just compared the output of the reference
implementation).
Also, I uncovered a fair amount of bugs in the selfhosted implementation
as part of this work, mainly related to variable shadowing within the
same function. I may need to implement some kind of more intelligent
variable resolution logic, rather than just assuming that the variable
will be available when emitting code to access it.