Selfhost: Compiling shadowed variable names
Within a function scope, a variable's name can be "shadowed" in child
scopes and it should be treated as a distinct variable from any
same-named variables in parent scopes. Previously, the selfhosted
compiler would naively keep track of variables in a function solely by
name, which of course doesn't allow for the following code, for example:
```
func foo() {
  val a = 1
  if true {
    val a = "hello"
  }
}
```
Instead, I added a mechanism for tracking and uniquely identifying a
variable within a function, indexing by the variable's name and
declaration location. At the time of compilation, every identifier is
known to be resolved to a variable, so as long as these variables are
indexed properly (using an incrementing int key to handle duplicates),
it is pretty easy to properly resolve variables in a function's body.