Selfhost: compiling lexer with selfhosted compiler! (#472)
* Selfhost: compiling lexer with selfhosted compiler!

I successfully compiled the selfhosted lexer (specifically the
`src/lexer.test.abra` file) and ran it with the test files in
`test/lexer/` to verify that the output was consistent when compared
with the output of running `src/lexer.test.abra` as compiled by the
reference compiler. In order to do this, I needed to refactor the test
runner a bit (and add some test output files which were previously
missing since it had just compared the output of the reference
implementation).
Also, I uncovered a fair amount of bugs in the selfhosted implementation
as part of this work, mainly related to variable shadowing within the
same function. I may need to implement some kind of more intelligent
variable resolution logic, rather than just assuming that the variable
will be available when emitting code to access it.

* Selfhost: Compiling shadowed variable names

Within a function scope, a variable's name can be "shadowed" in child
scopes and it should be treated as a distinct variable from any
same-named variables in parent scopes. Previously, the selfhosted
compiler would naively keep track of variables in a function solely by
name, which of course doesn't allow for the following code, for example:
```
func foo() {
  val a = 1
  if true {
    val a = "hello"
  }
}
```
Instead, I added a mechanism for tracking and uniquely identifying a
variable within a function, indexing by the variable's name and
declaration location. At the time of compilation, every identifier is
known to be resolved to a variable, so as long as these variables are
indexed properly (using an incrementing int key to handle duplicates),
it is pretty easy to properly resolve variables in a function's body.