Selfhost: compiling the parser
When compiling the `src/parser.test.abra` file, there were some
function invocations whose identifiers exceeded the length limit for qbe
identifiers. To fix this, I generate a randomized name to act as an
alias in place of that long type name. In order to perform the
randomization, I added an implementation of `rand` to the `libc` module
(and also added `srand(time(NULL))` to the `main` method. This is a
suboptimal solution, but it's good enough for now.
As of this change, the `src/parser.test.abra` file compiles using the
selfhosted compiler! However, that binary produces a segfault when run
on even a basic input, so I'll need ot dive in and debug.