Selfhost: compiling the parser
Attempting to compile and run the selfhosted parser using the
selfhosted compiler! There were some issues with the way that variables
were captured, and ultimately I solved the issue by making the whole
thing more performant anyway. This happens because rather than moving
every captured value to the heap when capturing it, I only move mutable
variables to the heap (they're the only types of variables that need to
have an extra layer of indirection applied, to handle reassignment).
Immutable variables (including parameters, which includes `self`) are
not moved to the heap, which means that the issues I had had trying to
run the selfhosted-compiled selfhosted parser are solved (there was a
closure that was trying to capture `self`, and that introduced errors).