Selfhost: compiling selfhosted parser
With this commit, the selfhosted compiler can compile the selfhosted
parser, and pass all of the tests that the reference-compiled
selfhosted parser passed!

One change I had to make in order to accomplish this was changing the
internal representation of Bools from being a 32-bit value to a 64-bit
value. This is due to the way that the garbage collector works, by
scanning allocated chunks for pointer-sized values that are also
themselves allocated chunks; chunks of memory that cannot be reached
in this manner are marked and swept. By representing boolean values
using 32 bits, values in types/enums were not aligned by a
pointer-sized amount, and the garbage collector was incorrectly
deallocating memory that was meant to be reachable, resulting in
undefined behavior.