Selfhost: compiling the parser (#473)
* Selfhost: compiling the parser

When compiling the `src/parser.test.abra` file, there were some
function invocations whose identifiers exceeded the length limit for qbe
identifiers. To fix this, I generate a randomized name to act as an
alias in place of that long type name. In order to perform the
randomization, I added an implementation of `rand` to the `libc` module
(and also added `srand(time(NULL))` to the `main` method. This is a
suboptimal solution, but it's good enough for now.
As of this change, the `src/parser.test.abra` file compiles using the
selfhosted compiler! However, that binary produces a segfault when run
on even a basic input, so I'll need ot dive in and debug.

* Selfhost: compiling the parser

Attempting to compile and run the selfhosted parser using the
selfhosted compiler! There were some issues with the way that variables
were captured, and ultimately I solved the issue by making the whole
thing more performant anyway. This happens because rather than moving
every captured value to the heap when capturing it, I only move mutable
variables to the heap (they're the only types of variables that need to
have an extra layer of indirection applied, to handle reassignment).
Immutable variables (including parameters, which includes `self`) are
not moved to the heap, which means that the issues I had had trying to
run the selfhosted-compiled selfhosted parser are solved (there was a
closure that was trying to capture `self`, and that introduced errors).

* Selfhost: compiling selfhosted parser

With this commit, the selfhosted compiler can compile the selfhosted
parser, and pass all of the tests that the reference-compiled
selfhosted parser passed!

One change I had to make in order to accomplish this was changing the
internal representation of Bools from being a 32-bit value to a 64-bit
value. This is due to the way that the garbage collector works, by
scanning allocated chunks for pointer-sized values that are also
themselves allocated chunks; chunks of memory that cannot be reached
in this manner are marked and swept. By representing boolean values
using 32 bits, values in types/enums were not aligned by a
pointer-sized amount, and the garbage collector was incorrectly
deallocating memory that was meant to be reachable, resulting in
undefined behavior.