Selfhost: compiling the typechecker
With this commit, the selfhosted compiler can compile the selfhosted
typechecker, and pass all of the tests that the reference-compiled
selfhosted typechecker passes!

Much like the prior commits, there were a few things I needed to fix up
that were only uncovered when attempting to compile the typechecker.
  1. I needed to fix up usages of the `getcwd` function, since
having `String` as a return type in the `libc` module is nonsensical (it
should be a `Pointer<Byte>` as that's the low-level C equivalent); also
callers of that function can, at the moment, just use the `buf` that's
passed as the first argument rather than using the return value.
  2. was the proper handling of terminators within while-loops in
functions (the function should only cease to be emitted when a
_returning_ terminator is seen, not just any terminator (eg. `break`)).
  3. Match expressions that have a type of `Never` (meaning that all of
the arms' blocks result in a terminator (for now) should just be handled
as if they were _statements_ instead.
  4. Lastly, ensure all destructuring has the correct arity (which is
now a requirement, unlike in the reference typechecker implementation).