Selfhost: compiling the typechecker
With this commit, the selfhosted compiler can compile the selfhosted
compiler, and pass all of the tests that the reference-compiled
selfhosted compiler passes! What a mouthful.

Anyway, the main change here was to not use the name `mem` for internal
locals after allocating memory, as that would interfere with parameters
named `mem` (what a niche corner case that was, but I'm glad it happened
and I was able to catch it; it took a while though). With this change, I
also updated the test runner to be able to run all of the compiler tests
against the compiler as compiled by itself, and they all pass! I'm very
proud of this accomplishment, as I've now implemented a fully selfhosted
compiler!