Selfhost: independent test runner
This adds a new test runner (in nodejs, for convenience) which uses
the current installation of `abra` to run the lexer, parser,
typechecker, and compiler tests against the assertion files. Once we
start using features in the compiler that are _not_ supported by the
reference implementation (for example, tuple destructuring, which is
already an example of the selfhosted compiler advancing beyond the
initial reference implementation), we will no longer rely on the
original rust-based implementation.