Selfhost: independent test runner (#477)
* Selfhost: independent test runner

This adds a new test runner (in nodejs, for convenience) which uses
the current installation of `abra` to run the lexer, parser,
typechecker, and compiler tests against the assertion files. Once we
start using features in the compiler that are _not_ supported by the
reference implementation (for example, tuple destructuring, which is
already an example of the selfhosted compiler advancing beyond the
initial reference implementation), we will no longer rely on the
original rust-based implementation.

* Tar flags...

* Tar flags...

* github action debugging

* github action debugging 2

* github action debugging 3

* github action debugging 4

* github action debugging 5

* github action debugging 6

* Cleanup, and adding colors to selfhost test runner

* Selfhost: adding tuple destructuring

Now that tuple destructuring is supported by the compiler (selfhosted)
and we're bootstrapping using the initial selfhosted compiler, update
the compiler to use tuple destructuring. Also, add a separate testing
pass to build the next compiler and ensure the tests pass against that
compiler too.

* Removing old rust-based selfhosted test suite

* Updating release gh actions script to not depend on reference rust+llvm compiler