Remove reference implementation (#478)
* Remove reference implementation

The selfhosted compiler has support for tuple destructuring, which is
now in use within the selfhosted compiler's source. As such, the
selfhosted compiler has progressed past beyond what the rust-based
reference compiler can support, and the reference compiler can now be
retired. This was always part of the selfhosted plan; now as new
language features and capabilities are supported by the selfhosted
compiler, the compiler can be re-generated to make use of these features
using a previous version to bootstrap.

* Adjust gh actions for new scripts dir