Better globals
Top-level variables in a module are only hoisted to globals if they're
referenced in a function (ie. captured). Since these variables can be
referenced as globals from within those functions, there's no need to
store them as captures for that function, and the function can
effectively be invoked as if it were an ordinary function.
Top-level variables that are _not_ referenced in functions are treated
effectively as locals within the module's wrapper function.
Exported variables (which can only be top-level) are always hoisted to
globals.