Keep track of callstack (#502)
Whenever a function is called, a 64-bit integer value is pushed onto an
"array" - this array is actually just a stack-allocated global of bytes
that is meant to emulate the callstack. The 64-bit integer represents an
encoding of the function being called, the file in which it was called,
and the line within that file on which it was called. After the function
returns, the value is "popped" from this array.