Const enum variants no longer backed by functions
Rather than having constant enum variants backed by a function call like
how container enum variants work, store them as constant `data` values
instead. Constant enum variants don't contain variables or data which
could vary depending on construction, so this is a pretty
straightforward (if minor) performance improvement.

The real main change hidden in this changeset is the solving of a sneaky
and nefarious bug that's been hidden in here for a while - when storing
Byte instances into a Pointer value using the intrinsic `Pointer#store`
method, it would actually use a `storel` qbe instruction behind the
scenes which would result in _4_ bytes being written to the underlying
memory. This had the evil side effect of overwriting the neighboring
memory segment's value, most likely to be `0`. This was most apparent in
`Array#toString` which I had actually already noticed a while ago but
never got around to actually digging in to figure out what was wrong.
This solution was the result of many hours' work (unfortunately) so it's
no wonder why I punted on it for so long. Oh well, it was fun.