LSP: Handling the 'initialize' method (#519)
* LSP: Handling the 'initialize' method

This adds a ton of code to the `json` std module to support json
encoding/decoding of the `initialize` request and response json rpc
messages as part of the LSP lifecycle. This can correctly receive an
`initialize` request and decode it to a data structure, and respond with
an `InitializeResult` object.

This code currently fails when it is sent the next message type (
`textDocument/didOpen`) which is not yet implemented.

* LSP: typecheck and emit diagnostics

Refactor the language server to be a bit more organized. Also, handle
the basic textDocument messages, for which the project is typechecked
and any errors are emitted as Diagnostics.