LSP: Using a virtual filesystem
Add a primitive notion of a "virtual filesystem" (which is just a
`Map<String, String>` for now) to represent abra files that require
typechecking but which have not yet been saved to disk. For now, the
contents of this map are wholesale replaced upon receipt of a
`textDocument/didChange` notification from the LSP client, and removed
upon `textDocument/didSave` (since at that point, the file can once
again be read from disk to obtain the latest version).

In the future, there ought to be a more sophisticated way of
representing the file contents in the virtual filesystem, which
leverages the Incremental `contentChange` option for
`textDocument/didChange` notifications, rather than having to wastefully
parse and replace the entire contents of the document each time. In
fact, the LSP currently falls over when trying to operate on large
files, most likely for this reason.