Postponing virtual filesystem changes
Using this current approach, the entire contents of the changed file
will be sent in `textDocument/didChange` events (as well as
`textDocument/didOpen`, which is less relevant but still wasteful).
Until a more performant system is developed to leverage Incremental
didChange events, let's only run diagnostics on save and instruct the
client to not send didOpen or didChange events. Additionally, we
instruct the client to not send the file contents in the didSave event.