LSP: Go to definition (#526)
* LSP: Go to definition

Implement goto definition in the language service, for variables and
standalone identifiers (much like with hover, there's still a lot of
places that need to have identifier metadata emitted during
typechecking). As those places are updated, they will work for hover as
well as goto definition, which is a big upside!
This also refactors the `Some(...)` AST transformation to use a bogus
Position, so hover and goto definition work for `Some(...)` identifiers.

* Fixing tests since Option.Some and Option.None transformations now use bogus position