LSP: cleanup, refactoring, more idents (#534)
* LSP: cleanup, refactoring, more idents

Adds ident support (hover, goto definition) for enum variants and
methods/fields. I also refactored some stuff to make it a bit more
organized since the existing LSP-specific logic was getting a little out
of control. There's definitely more cleanup work to be done, but this is
just a good first pass.

* LSP: Idents for match cases

This also changes the Parser transformation for `Ok` and `Err` as match
cases to use `Position.bogus()` rather than the same position as the
other token.