Cleanup: improve invocation codegen
Previously there was a lot of logic involved in the compiler when
performing codegen for invocation expressions, particularly in the case
of invoking an expression value (aka, a function value). This change
abstracts that complex logic out into a generated `#call` method for the
adhoc "Function" structs, and it also cleans it up a bunch too.
Additionally, the logic in the main codegen function for invocation
expressions gets significantly cleaner.