Introduce `pub` keyword
The first pass is to replace the `export` keyword with `pub`. This saves
from having two keywords (`export` for use at the top-level, and `pub`
for type fields/methods) and instead the usage of `pub` at the
top-level indicates that the value is exported (aka, visible outside of
the current module, similarly to how `pub` fields/methods are visible
outside of the current type/enum).