Introduce `pub` keyword (#542)
* Introduce `pub` keyword

The first pass is to replace the `export` keyword with `pub`. This saves
from having two keywords (`export` for use at the top-level, and `pub`
for type fields/methods) and instead the usage of `pub` at the
top-level indicates that the value is exported (aka, visible outside of
the current module, similarly to how `pub` fields/methods are visible
outside of the current type/enum).

* Typechecker: error changes + cleanup

Clean up the typechecker by tracking the current TypedModule rather than
tracking individually all the useful pieces. This just seems like a
cleaner way of doing things. Also, since the `pub` keyword is being
introduced and the `export` keyword phased out, adjust the error
messages related to exports to better reflect that new world.

* Parser/Typechecker: Support `pub` fields/methods

This doesn't actually enforce visibility on fields or methods, but it
does track them. This was a large change which required lots of test
updates.

* Fixing failing test