Enforcing `pub` visibility for fields
Ensure that a non-`pub` field on a type cannot be accessed outside of
the module (file) in which it was defined.

This involved a lot of changes across a ton of files in order to
maintain compliance. There will be even more things that change once
methods' visibility with `pub` are also enforced.