Enforcing `pub` visibility for fields (#544)
* Enforcing `pub` visibility for fields

Ensure that a non-`pub` field on a type cannot be accessed outside of
the module (file) in which it was defined.

This involved a lot of changes across a ton of files in order to
maintain compliance. There will be even more things that change once
methods' visibility with `pub` are also enforced.

* Allowing non-pub access within std

For performance reasons, don't require modules within std/ to use
wrapper methods in order to access non-pub fields on standard library
types.