LSP: Handle `textDocument/documentSymbol`
This change allows the LSP to provide a list of symbols for the current
document, which includes a nested hierarchical view of types and their
fields, enums and their variants, instance and static methods for each,
as well as top-level variables and functions.

Doing this caused a few bugs and gaps to surface, which are also fixed
as part of this changeset. These include:

Lexer: Consider a \t as a whitespace character when previously it
wasn't.

Typechecker: When adding a TypedModule to the Project, make sure to use
the normalized path name, with `/a/b/../c` resolved to `/a/c`
(previously this wasn't the case).

Language service: If the TypedModule isn't currently present in the
project, pre-emptively typecheck the current module rather than waiting
for a save. This is helpful because it allows for more seamless
go-to-definition and hover commands, but also because I think if the
symbols aren't resolved correctly when the file first opens, vscode will
ignore subsequent symbols provided by `textDocument/documentSymbol`. I
also cleaned up residual "virtual file system" changes which had been
commented out but which can now be disregarded because I don't think I'm
going to be doing that any time soon.