Succeed typechecking but with errors
Previously, if typechecking encountered a type error (or an error
reading, lexing, or parsing the file, which it packaged up as a
typechecker error), it exited the program and surfaced the error. This
works great, however it only allows one error to be surfaced at a time,
and it also halts typechecking for the rest of the file and prevents any
other meaningful type information to be gathered. This notably includes
data for the LSP, so if typechecking failed there would be no
go-to-definition/hover support and no symbol support, until all of the
type errors were addressed (which could be cumbersome, if there are more
than one and they're only revealed one at a time).

Taking this approach, wherein typechecking is allowed to recover from
errors rather than bailing early (and keeping track of those errors to
be surfaced later), allows the module to be at least partially typed.
This, in turn, allows the LSP to still provide functionality.