Adding `CouldNotDetermine` type
This type will be used as the type of any expression that has an
associated typechecking error. It's assignable to and from any type, to
allow for as much code as possible to be visited by the typechecker.
This change works through a few of the typechecking handlers for various
expressions, but eventually they'll have to be covered.

This also adds a test file for this particular use case, which has two
assertion files; one assertion file verifies the error messages that are
raised and the other validates that the `CouldNotDetermine` type makes
its way through the typed expressions of the module, and also that the
full module gets emitted correctly, which would not be the case if an
unrecoverable error were encountered (the removal of which is the goal
of all of this work).