Officially adding syntax for decorators (#553)
Step 0 is a cleanup of existing decorators. Currently there's
`@Intrinsic`, `@CBinding`, and `@Stub` which are used throughout the
compiler and standard library. These can be consolidated and reformatted
to fit the eventual future design ideas. For one, `@Stub` does not need
to exist now that `todo` exists. Additionally since `@Intrinsic` and
`@CBinding` really are more like language-level keywords, they should be
lowercased (and `@CBinding` is being renamed to something more
self-explanatory - `@external`).