Lexer & Parser: Adding support for `decorator` (#558)
This is take 2 of this change, as I was unhappy with the way I did it
the first time around. Basically, this time I decided to implement
decorators as just being syntactic wrapper on top of existing `type`
declarations. They'll have all the same semantics when it comes to
typechecking, and implementing them in terms of existing type
declarations saves from having a lot of duplicate logic.