Cleanup
Removing the special handling logic for `print`/`println`, and also roll
back the standalone generic function resolution fix. This fix is now
just removing instances of `print`/`println` and replacing it with a
(much less elegant) call to `stdoutWrite`/`stdoutWriteln`. In the
future, once the IR implementation is done and the generic resolution
algorithm is rewritten (and simplified), it may make sense to re-add the
`print`/`println` functions (possibly as recipients of an
as-yet-unspec'ed "variadic generics" type?).
Note also that `print`/`println` are 2 of the 3 recipients of the `Any`
type, which can also be deprecated.