Ensure functions are fully genericized (#565)
* Ensure functions are fully genericized

This addresses #516 and adds generic resolution to top-level functions,
as well as methods that have generics other than their parent type's.

* Back to print/println

I had renamed everything to make it easier to refactor, and now I'm
renaming them back.

* Moving off of print/println temporarily

* Add stdoutWriteln function

* Whoops lol

* I missed some 'print' calls

* Fixing the callstack test

* Don't update existing print/println functions

* Cleanup

Removing the special handling logic for `print`/`println`, and also roll
back the standalone generic function resolution fix. This fix is now
just removing instances of `print`/`println` and replacing it with a
(much less elegant) call to `stdoutWrite`/`stdoutWriteln`. In the
future, once the IR implementation is done and the generic resolution
algorithm is rewritten (and simplified), it may make sense to re-add the
`print`/`println` functions (possibly as recipients of an
as-yet-unspec'ed "variadic generics" type?).
Note also that `print`/`println` are 2 of the 3 recipients of the `Any`
type, which can also be deprecated.