Removing `Any` type
The time has come. I've decided to not have a dedicated notion of
"traits" or interfaces in the language, instead on just using types and
passing required functions as values. In the future, perhaps this can be
simplified with compiletime functions, but for now I think that (plus
generics) provide enough of a surface area that traits/interfaces will
not be required.
As such, it's time to remove the primordial trait/interface which has
been a part of the language since the beginning: `Any`. There is now no
such thing as `Any`. The notion of "a type for which the only
information we have is that it has toString/hash/eq" is effectively
equivalent to a generic `T`; in fact `println` can (non-variadically) be
implemented as just:
```
func println<T>(item: T) = stdoutWriteln(item.toString())
```
In the future, some notion of variadic generics may be added which would
allow for something like this:
```
func println<*T>(*items: T[]) {
  for item in items {
    stdoutWrite(item.toString())
    stdoutWrite(" ")
  }
  stdoutWrite("\n")
}
```
However, that syntax doesn't exist yet (and may not ever, who knows?).
But for now, let's get rid of the last vestiges of `Any` and do some
minor cleanup.