First meaningful chunk of IR generation
The IR generator can now generate IR for function and method bodies as
well as invocation of functions and methods. It also handles basic
struct initializers and field accessing. I'm hopeful that the
introduction of `ConcreteType` during IR generation will allow the IR
_compilation_ (to QBE) to be pretty small since the brunt of the work
figuring out types and generics and whatnot will be done during IR
generation.
Still remaining to be done is an interpreter of IR code. I think that
will come next, since that's the whole goal of this exercise and I want
to ensure that the IR "language" evolves in step with native binary
compilation and a future VM, since eventually this will be used to power
compile-time evaluation of code (eg. macros, etc).