IR: Working towards printing an array
This introduces the notion of "concrete generics" which is the
next-generation of the "resolved generics" idea from the previous
compiler but it uses `ConcreteType`s instead (which are vastly easier to
work with). Doing so paves the way for compile-time generic code
generation of generic functions (standalone, and instance/static
methods).
This also adds in a few improvements and streamlining of variable/temp
creation.