IR: creating array literals
The main objective for this change was to successfully generate and
compile IR for the following code, and have the compiled code execute
correctly:
```
var arr = [1, 2, 3]
stdoutWriteln(arr.length.toString())

arr.push(4)
arr.push(5)
arr.push(6)
stdoutWriteln(arr.length.toString())
```

In order to accomplish this, a lot of changes had to happen.
- IR/Compiler: This adds a ton of new functionality and instructions,
including new builtins, arithmetic operators, if stmt/expr, equality,
and more.
- Typechecker: The way I was doing generic resolution was incorrect and
clunky for instance methods which reference their parent's type
parameters. By switching to resolving those generics at the call-site
rather than the accessor-site, I'm able to track that as a resolved
generic which can be used in later IR generation. This was a relatively
small change to the typechecker in terms of lines of code, but it was a
fairly large cognitive shift I had to make in order to think about
generic resolution differently, since that's probably the most complex
part of the typechecker as it is today.
An additional typechecker change was tracking the number of child scopes
within a function, so variables in child scopes which shadow names from
upper scopes can be generated efficiently (by adding a suffix based on
the scope index within the function).