First meaningful chunk of IR generation (#567)
* First meaningful chunk of IR generation

The IR generator can now generate IR for function and method bodies as
well as invocation of functions and methods. It also handles basic
struct initializers and field accessing. I'm hopeful that the
introduction of `ConcreteType` during IR generation will allow the IR
_compilation_ (to QBE) to be pretty small since the brunt of the work
figuring out types and generics and whatnot will be done during IR
generation.
Still remaining to be done is an interpreter of IR code. I think that
will come next, since that's the whole goal of this exercise and I want
to ensure that the IR "language" evolves in step with native binary
compilation and a future VM, since eventually this will be used to power
compile-time evaluation of code (eg. macros, etc).

* Fixing that process_callstack test again

* Compiler optimizations

Add basic constant folding for adding constant ints, as well as constant
global initialization. This also lifts constant strings up into the data
section (the whole struct contents is embedded in memory) which saves
from unnecessary runtime memory allocations.

* Extract out raw qbe ssa code for builtin functions

* IR: Working towards printing an array

This introduces the notion of "concrete generics" which is the
next-generation of the "resolved generics" idea from the previous
compiler but it uses `ConcreteType`s instead (which are vastly easier to
work with). Doing so paves the way for compile-time generic code
generation of generic functions (standalone, and instance/static
methods).
This also adds in a few improvements and streamlining of variable/temp
creation.

* mutable variables should be stack-allocated, in case they're reassigned to

* IR: creating array literals

The main objective for this change was to successfully generate and
compile IR for the following code, and have the compiled code execute
correctly:
```
var arr = [1, 2, 3]
stdoutWriteln(arr.length.toString())

arr.push(4)
arr.push(5)
arr.push(6)
stdoutWriteln(arr.length.toString())
```

In order to accomplish this, a lot of changes had to happen.
- IR/Compiler: This adds a ton of new functionality and instructions,
including new builtins, arithmetic operators, if stmt/expr, equality,
and more.
- Typechecker: The way I was doing generic resolution was incorrect and
clunky for instance methods which reference their parent's type
parameters. By switching to resolving those generics at the call-site
rather than the accessor-site, I'm able to track that as a resolved
generic which can be used in later IR generation. This was a relatively
small change to the typechecker in terms of lines of code, but it was a
fairly large cognitive shift I had to make in order to think about
generic resolution differently, since that's probably the most complex
part of the typechecker as it is today.
An additional typechecker change was tracking the number of child scopes
within a function, so variables in child scopes which shadow names from
upper scopes can be generated efficiently (by adding a suffix based on
the scope index within the function).