IR: compile to javascript
One of the goals of introducing the IR is to allow for compilation to
different targets. As I continue to develop and expand upon the IR, I
wanted to introduce a secondary compilation target (javascript) as a
means of ensuring that the representation is flexible enough to support
more than just native compilation (via qbe).
As a result of this, I've expanded the Load and Store operations to have
LoadField and StoreField, with the possible intent of replacing Load and
Store with the LoadPtr/StorePtr builtins (since they should at that
point only be used with raw pointer manipulation), but that remains to
be seen. Anyway, this change successfully compiles the generated IR into
a javascript module which is _almost_ runnable (either via node or
within a browser). The one thing that's missing is the raw pointer
load/store operations, which will need some reimagining in order to
support. This may end up being a good thing though, since it allows for
some cleanup of the intrinsic `Pointer` methods, which can be a bit
clunky to work with (even though they're not meant to be used often).