Improved pointer intrinsic methods (part 1/2)
While working on IR compilation for the js target, I realized that the
existing Pointer intrinsics for store/load weren't sufficient because
the potential offset is calculated ahead of time. While there may have
been a way to successfully convert the IR into javascript code, it's
also true that the existing way of interacting with these low-level
Pointer methods is sort of clunky.