Use updated `store_at` semantics
All the work over the past several changes was to enable tracking the
offset in the Store builtin (aka `pointer_store_at`) to more easily
transform the IR into javascript. It also simplifies the native compiler
by removing the `pointer_offset` builtin, since now that's baked into
`pointer_store_at` and `pointer_load_at`. Arguably the existing Load and
Store operations should be replaced with LoadVar and StoreVar since
that's the only time that they're actually used now, but that can be
another changeset if I so choose.
As a proof of concept of js code generation from the IR, I have two new
example files: `example.mjs` (meant to run via `node`), and
`example.html` (meant to be run in browser). Both of these work, and
each uses different semantics for the `write` extern (outputting to the
terminal vs appending to a DOM element).